# internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: Social API
  version: "1.0.0"
  summary: Social posts with an optional image and a global chronological feed.
  description: |
    Implementation of the contract in `HANDOFF.md`.

    ### Conventions

    - JSON only, `camelCase` in every request and response body.
    - Timestamps are ISO 8601 UTC with exactly three fractional digits:
      `2026-09-04T11:23:00.000Z`.
    - IDs and page cursors are **opaque strings**. Never parse, sort, or do
      arithmetic on them; pass a cursor back exactly as it was received.
    - Successful responses return the resource directly. The only envelope is on
      paginated lists, which return `{ items, nextCursor }`.
    - User-facing copy is British English.

    ### Authentication

    Two tokens. The access token lasts 15 minutes and is sent on every protected
    request as `Authorization: Bearer <accessToken>`. The refresh token lasts 30
    days and is only ever sent to `/v1/auth/refresh`.

    Refresh tokens **rotate**: every call to `/v1/auth/refresh` returns a new
    refresh token and invalidates the one presented. Persist the new one
    immediately. Presenting an already-used refresh token is treated as a leak
    and revokes every token descended from that login.

    ### Images

    Image bytes never pass through this API. Ask for a presigned URL, `PUT` the
    bytes straight to object storage with **no `Authorization` header of ours**,
    then send the returned `key` back on the resource. See
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /v1/auth/refresh:
    post:
      tags: [Authentication]
      summary: Exchange a refresh token for a new pair
      operationId: refresh
      security: []
      description: |
        Public, because the refresh token is itself the credential.

        This rotates: the response carries a **new** refresh token and the one
        you sent is now spent. Persist the new one before doing anything else.

        Refresh must be single-flight on the client. If several requests get a
        401 at once, queue them behind one refresh call — a second call using
        the already-rotated token will fail and revoke the whole session.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RefreshTokenRequest" }
      responses:
        "200":
          description: A new access and refresh token pair.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResponse" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "401":
          description: |
            The refresh token is unknown, expired, already used, or revoked.
            There is nothing left to refresh with: clear both tokens and send the
            user to the sign-in screen.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example:
                error:
                  code: TOKEN_INVALID
                  message: Your session is no longer valid. Please sign in again.
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        The `accessToken` from register, login or refresh, sent as
        `Authorization: Bearer <accessToken>`. Valid for 15 minutes.

        A 401 carries one of two codes, and they mean different things:
        `TOKEN_EXPIRED` means refresh and retry the request once;
        `TOKEN_INVALID` means log the user out.
...
  responses:
    ValidationFailed:
      description: |
        The body failed validation. `details` maps each rejected field to a
        reason; use it to drive inline form errors.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error:
              code: VALIDATION_FAILED
              message: Please check the highlighted fields.
              details:
                username: may only contain lowercase letters, numbers and underscores
                password: must be at least 8 characters
...
    RateLimited:
      description: Too many requests. Back off and honour `Retry-After`.
      headers:
        Retry-After:
          description: Whole seconds to wait before retrying.
          schema: { type: integer, minimum: 1 }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error:
              code: RATE_LIMITED
              message: Too many requests. Please try again shortly.
    InternalError:
      description: |
        Something failed on the server. Show a generic message and log the rest;
        the response deliberately carries no internal detail.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error:
              code: INTERNAL_ERROR
              message: Something went wrong on our end. Please try again.
  schemas:
    Timestamp:
      type: string
      format: date-time
      description: ISO 8601 UTC with exactly three fractional digits.
      pattern: '^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}\.\d{3}Z$'
      examples: ["2026-09-04T11:23:00.000Z"]
    Username:
      type: string
      minLength: 3
      maxLength: 30
      pattern: "^[a-z0-9_]+$"
      description: Lowercase letters, digits and underscores. Immutable after registration.
      examples: [ada]
...
    Error:
      type: object
      description: |
        The shape of every failure. `message` is safe to show to the user;
        `code` never is.
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Stable identifier to branch on. Never render this.
              enum:
                - VALIDATION_FAILED
                - INVALID_CREDENTIALS
                - TOKEN_EXPIRED
                - TOKEN_INVALID
                - FORBIDDEN
                - NOT_FOUND
                - CONFLICT
                - FILE_TOO_LARGE
                - UNSUPPORTED_MEDIA_TYPE
                - RATE_LIMITED
                - INTERNAL_ERROR
            message:
              type: string
              description: User-facing British English. Safe to display.
            details:
              type: object
              description: Field name to reason. Present on validation and conflict failures.
              additionalProperties: { type: string }
          required: [code, message]
      required: [error]
    User:
      type: object
      description: The full private profile. Only ever returned to its owner.
      properties:
        id: { type: string, examples: [djbaalwbrctr35h4bmtjjald5y] }
        email: { type: string, format: email, maxLength: 254, examples: ["ada@example.com"] }
        username: { $ref: "#/components/schemas/Username" }
        displayName: { type: string, minLength: 1, maxLength: 50, examples: [Ada Lovelace] }
        bio: { type: string, maxLength: 160, examples: ["Building things."] }
        avatarUrl:
          type: [string, "null"]
          format: uri
          description: Null when the user has no avatar.
        createdAt: { $ref: "#/components/schemas/Timestamp" }
      required: [id, email, username, displayName, bio, avatarUrl, createdAt]
...
    AuthResponse:
      type: object
      description: |
        Returned by register, login and refresh. Carries the user so no follow-up
        request is needed, and the access token expiry so the client can refresh
        before a request fails rather than after.
      properties:
        accessToken:
          type: string
          description: "Send as `Authorization: Bearer …`. Hold in memory only."
        refreshToken:
          type: string
          description: |
            Store in secure device storage, never in plain async storage. Replaced
            on every refresh — persist the new one immediately.
        accessTokenExpiresAt: { $ref: "#/components/schemas/Timestamp" }
        user: { $ref: "#/components/schemas/User" }
      required: [accessToken, refreshToken, accessTokenExpiresAt, user]
...
    RefreshTokenRequest:
      type: object
      properties:
        refreshToken: { type: string }
      required: [refreshToken]
...